+++ ruby/build_templates.rb
@@ -6,25 +6,38 @@
 # target: <app.rb>
 module Hokusai
   def self.docker_template
-    <<~EOF
+    <<~HELL
 FROM skinnyjames/mruby-cross-<%= os %> as cross
     
 RUN apt update -y && apt-get install -y wget <%= deps %>
 
 WORKDIR /temp
-RUN wget https://github.com/skinnyjames/mruby-bin-barista/releases/download/0.2.4/barista-linux-x86.tar.gz && \
+RUN wget https://github.com/skinnyjames/mruby-bin-barista/releases/download/0.3.1/barista-linux-x86.tar.gz && \
     tar -xvf barista-linux-x86.tar.gz && \
-    chmod 755 barista-linux/barista && \
-    cp barista-linux/barista /usr/bin/.
+    chmod 755 barista-linux-x86/barista && \
+    cp barista-linux-x86/barista /usr/bin/.
 
 WORKDIR /app
 
 RUN git clone --branch 5.5 --depth 1 https://github.com/raysan5/raylib.git vendor/raylib
 RUN git clone --depth 1 https://github.com/tree-sitter/tree-sitter.git vendor/tree-sitter
 RUN git clone --branch stable --depth 1 https://github.com/mruby/mruby.git vendor/mruby
-RUN git clone --branch main --depth 1 https://github.com/skinnyjames/hokusai-pocket.git vendor/hp
+RUN git clone --branch feature/networking --depth 1 https://github.com/skinnyjames/hokusai-pocket.git vendor/hp
 RUN git clone https://github.com/mlabbe/nativefiledialog.git vendor/nfd
 RUN git clone https://github.com/libuv/libuv vendor/libuv
+
+# fetch http deps
+RUN wget -O vendor/llhttp.tar.gz https://github.com/nodejs/llhttp/archive/refs/tags/release/v9.3.1.tar.gz && \
+    cd vendor && tar -xvf llhttp.tar.gz && mv llhttp-release-v9.3.1 llhttp
+
+RUN wget -O vendor/mbedtls.tar.bz2 https://github.com/Mbed-TLS/mbedtls/releases/download/mbedtls-3.6.6/mbedtls-3.6.6.tar.bz2 && \
+    cd vendor && tar -xvf mbedtls.tar.bz2 && mv mbedtls-3.6.6 mbedtls
+
+RUN git clone https://github.com/madler/zlib.git vendor/zlib
+
+RUN wget -O vendor/tlsuv.tar.gz https://github.com/openziti/tlsuv/archive/refs/tags/v0.41.1.tar.gz && \
+    cd vendor && tar -xvf tlsuv.tar.gz && mv tlsuv-0.41.1 tlsuv
+    
 
 # build mruby
 WORKDIR /app/vendor/mruby
@@ -172,15 +185,40 @@ index a26b8ce6..798d7bd0 100644
  RLAPI void ImageDither(Image *image, int rBpp, int gBpp, int bBpp, int aBpp);                            // Dither image data to 16bpp or lower (Floyd-Steinberg dithering)
 EOT
 
+RUN apt install -y gpg
+
+# Source - https://stackoverflow.com/a/56690743
+# Posted by Liu Hao Cheng, modified by community. See post 'Timeline' for change history
+# Retrieved 2026-05-18, License - CC BY-SA 4.0
+RUN wget -O - https://apt.kitware.com/keys/kitware-archive-latest.asc 2>/dev/null | gpg --dearmor - | tee /usr/share/keyrings/kitware-archive-keyring.gpg >/dev/null
+RUN echo 'deb [signed-by=/usr/share/keyrings/kitware-archive-keyring.gpg] https://apt.kitware.com/ubuntu/ jammy main' | tee /etc/apt/sources.list.d/kitware.list >/dev/null
+RUN apt-get update -y && apt-get install -y cmake
+
 <% if os == "windows" %>
+# Create the mingw64-cmake wrapper
+RUN echo '#!/bin/sh\\nexec cmake -DCMAKE_SYSTEM_NAME=Windows -DCMAKE_C_COMPILER=x86_64-w64-mingw32-gcc -DCMAKE_CXX_COMPILER=x86_64-w64-mingw32-g++ -DCMAKE_FIND_ROOT_PATH=/usr/x86_64-w64-mingw32 "$@"' > /usr/local/bin/cmake-wrap \
+    && chmod +x /usr/local/bin/cmake-wrap
 ENV CC=x86_64-w64-mingw32-gcc-posix
 ENV AR=x86_64-w64-mingw32-gcc-ar
+ENV ZLIBA="libzs.a"
 <% elsif os == "osx" %>
+RUN echo '#!/bin/sh\\nexec cmake -DCMAKE_SYSTEM_NAME=Darwin -DCMAKE_C_COMPILER=x86_64-apple-darwin20.4-clang -DCMAKE_CXX_COMPILER=x86_64-apple-darwin20.4-clang++ -DCMAKE_FIND_ROOT_PATH=/opt/osxcross/target "$@"' > /usr/local/bin/cmake-wrap \
+    && chmod +x /usr/local/bin/cmake-wrap
+ENV OSXCROSS_ROOT=/opt/osxcross/target
+ENV OSXCROSS_HOST=x86_64-linux-gnu
+ENV OSXCROSS_TARGET_DIR=/opt/osxcross/target
+ENV OSXCROSS_TARGET=x86_64-apple-darwin20.4
+ENV OSXCROSS_SDK_DIR=$OSXCROSS_TARGET_DIR/SDK
+ENV OSXCROSS_SDK="MacOSX11.3.sdk"
 ENV CC=x86_64-apple-darwin20.4-clang
 ENV AR=x86_64-apple-darwin20.4-ar
+ENV ZLIBA="libz.a"
 <% else %>
+RUN echo '#!/bin/sh\\nexec cmake "$@"' > /usr/local/bin/cmake-wrap \
+    && chmod +x /usr/local/bin/cmake-wrap
 ENV CC=gcc
 ENV AR=ar
+ENV ZLIBA="libz.a"
 <% end %>
 
 WORKDIR /app/vendor/raylib
@@ -218,16 +256,62 @@ RUN cd build/gmake_macosx && make config=release_x64
 RUN cd build/gmake_linux_zenity && make config=release_x64
 <% end %>
 
+
 # build libuv
 WORKDIR /app/vendor/libuv
-RUN apt update -y && apt install -y automake libtool
-<% if os == "osx" %>
-RUN ./autogen.sh && ./configure --host=x86_64-apple-darwin20.4 && make
-<% elsif os == "windows" %>
-RUN ./autogen.sh && ./configure --host=x86_64-w64-mingw32 && make
-<% else %>
-RUN ./autogen.sh && ./configure && make
-<% end %>
+RUN cmake-wrap -S . -B build -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_LIBDIR=lib -DCMAKE_INSTALL_PREFIX=build/dist
+RUN cd build && make -j 5 all install
+
+# build llhttp
+WORKDIR /app/vendor/llhttp
+RUN mkdir -p build
+RUN mkdir -p dist
+RUN cmake-wrap -S . -B build -DCMAKE_BUILD_TYPE=Release -DLLHTTP_BUILD_STATIC_LIBS=ON -DCMAKE_INSTALL_LIBDIR=lib -DCMAKE_INSTALL_PREFIX=dist
+RUN cd build && make -j 5 install
+
+# build mbedtls
+WORKDIR /app/vendor/mbedtls
+RUN mkdir -p build
+RUN cmake-wrap -S . -B build -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=build/dist -DCMAKE_INSTALL_LIBDIR=lib -DENABLE_TESTING=OFF -DENABLE_PROGRAMS=OFF
+RUN cd build && make -j 5 install
+
+# build zlib
+WORKDIR /app/vendor/zlib
+RUN mkdir -p build
+RUN cmake-wrap -S . -B build -DCMAKE_BUILD_TYPE=Release -DZLIB_BUILD_TESTING=OFF -DZLIB_BUILD_SHARED=OFF -DZLIB_INSTALL=OFF
+RUN cd build && make -j 5
+
+# build tlsuv
+WORKDIR /app/vendor/tlsuv
+COPY <<'FUCK' tlsuv.patch
+#{Hokusai::Patches.tlsuv_patch}
+FUCK
+
+RUN git init
+RUN git add .
+RUN git apply tlsuv.patch
+
+RUN cmake-wrap -S . -B build DMBEDCRYPTO_LIBRARY='../../vendor/mbedtls/build/dist/libmbedcrypto.a' \
+  -DMBEDTLS_INCLUDE_DIRS='../../vendor/mbedtls/build/dist/include' \
+  -DMBEDTLS_LIBRARY='../../vendor/mbedtls/build/dist/lib/libmbedtls.a' \
+  -DMBEDX509_LIBRARY='../../vendor/mbedtls/build/dist/lib/libmbedx509.a' \
+  -DBUILD_SHARED_LIBS=OFF \
+  -DCMAKE_BUILD_TYPE=Release \
+  -DTLSUV_HTTP=ON \
+  -DTLSUV_TLSLIB=mbedtls \
+  -DZLIB_INCLUDE='../../vendor/zlib' \
+  -DZLIB_LIB="../../vendor/zlib/build/$ZLIBA" \
+  -DLLHTTP_LIB='../../vendor/llhtp/dist/lib/libllhttp.a' \
+  -DLLHTTP_INCLUDE='../../vendor/llhttp/dist/include' \
+  -DTLSUV_LIBUV_LIB='../../vendor/libuv/libuv.a' \
+  -DTLSUV_LIBUV_INCLUDE='../../vendor/libuv/build/dist/include' \
+<% if os == "osx" %>\
+  -DCMAKE_FRAMEWORK_PATH='/opt/osxcross/target/SDK/MacOSX11.3.sdk/System/Library/Frameworks' \
+  -DCMAKE_EXE_LINKER_FLAGS='-framework Security'\
+<% end %>\
+  -DMBEDTLS_INCLUDE='../../vendor/mbedtls/build/dist/include/' 
+
+RUN cd build && make -j 5 all 
 
 WORKDIR /app
 RUN mkdir -p /app/vendor/hokusai-pocket
@@ -240,10 +324,18 @@ spec("hokusai-pocket-app") do
     end
 
 <% if os.eql?("windows")%>
+    def zlib
+      "libzs.a"
+    end
+
     def nfd
       "nfd.lib"
     end
 <% else %>
+    def zlib
+      "libz.a"
+    end
+
     def nfd
       "libnfd.a"
     end
@@ -251,11 +343,11 @@ spec("hokusai-pocket-app") do
 
 <% if os.eql?("windows") %>
     def libs
-      "-lws2_32 -lgdi32 -lwinmm -lcomctl32 -lcomdlg32 -lole32 -luuid -ldbghelp -liphlpapi -luserenv"
+      "-lws2_32 -lgdi32 -lwinmm -lcomctl32 -lcomdlg32 -lole32 -luuid -ldbghelp -liphlpapi -luserenv -lbcrypt -lcrypt32 -static -lwinpthread"
     end
 <% elsif os.eql?("osx") %>
     def libs
-      "-framework CoreVideo -framework CoreAudio -framework AppKit -framework IOKit -framework Cocoa -framework GLUT -framework OpenGL"
+      "-framework CoreVideo -framework Security -framework CoreAudio -framework AppKit -framework IOKit -framework Cocoa -framework GLUT -framework OpenGL"
     end
 <% else %>
     def libs
@@ -267,24 +359,45 @@ spec("hokusai-pocket-app") do
           vendor/tree-sitter/build/include 
           vendor/raylib/src 
           vendor/mruby/include
+          vendor/mruby/build/host/include
           vendor/hp/grammar/tree_sitter
           vendor/hp/src
           vendor/hp/src/mruby-uv
           vendor/nfd/src/include
           vendor/libuv/include
+          vendor/llhttp/include
+          vendor/tlsuv/deps/uv_link_t/include
+          vendor/tlsuv/build/generated
+          vendor/tlsuv/include
+          vendor/zlib
+          vendor/hp/src/http
         ]
     end
 
+    def mbedtls_libs
+      %w[libmbedtls.a libmbedx509.a libmbedcrypto.a]
+    end
+
     def links
-      (%w[
+      ln = %w[
         vendor/hp/grammar/src/parser.c
         vendor/hp/grammar/src/scanner.c
         vendor/hokusai-pocket/libhokusai.a
         vendor/mruby/build/platform/lib/libmruby.a 
         vendor/raylib/src/libraylib.a
         vendor/tree-sitter/build/lib/libtree-sitter.a
-        vendor/libuv/.libs/libuv.a
-      ] + ["vendor/nfd/build/lib/Release/x64/\#{nfd}"]).join(" ")
+        vendor/libuv/build/dist/lib/libuv.a
+      ] + ["vendor/nfd/build/lib/Release/x64/\#{nfd}"]
+
+      ln << "vendor/tlsuv/build/libtlsuv.a"
+      ln << "vendor/llhttp/dist/lib/libllhttp.a"
+
+      mbedtls_libs.each do |lib|
+        ln << "vendor/mbedtls/build/dist/lib/\#{lib}"
+      end
+
+      ln << "vendor/zlib/build/\#{zlib}"
+      ln.join(" ")
     end
 
     def h_includes
@@ -340,6 +453,7 @@ spec("hokusai-pocket-app") do
       # ugh, need separate libuv/raylib compilation units because of windows.h collisions
       loop_includes = %w[
         vendor/mruby/include
+        vendor/mruby/build/host/include
         vendor/libuv/include
         vendor/tree-sitter/build/include
         vendor/hp/src
@@ -402,6 +516,7 @@ spec("hokusai-pocket-app") do
         vendor/raylib/src
         vendor/tree-sitter/build/include 
         vendor/mruby/include
+        vendor/mruby/build/host/include
         .
         vendor/hokusai-pocket
         vendor/hp/src
@@ -437,6 +552,6 @@ RUN barista build
 # export
 FROM scratch
 COPY --from=cross /app/bin/ /<%= outfile %>
-EOF
+HELL
   end
 end